requirements.txt
django-cors-headers
streamlit
requests